moved notifier to service package top level
