+++ service/deployer/deployer.go
@@ -10,11 +10,11 @@ import (
 	"github.com/giantswarm/draughtsman/flag"
 	"github.com/giantswarm/draughtsman/service/deployer/installer"
 	installerspec "github.com/giantswarm/draughtsman/service/deployer/installer/spec"
-	"github.com/giantswarm/draughtsman/service/deployer/notifier"
-	notifierspec "github.com/giantswarm/draughtsman/service/deployer/notifier/spec"
 	"github.com/giantswarm/draughtsman/service/eventer"
 	eventerspec "github.com/giantswarm/draughtsman/service/eventer/spec"
 	httpspec "github.com/giantswarm/draughtsman/service/http"
+	"github.com/giantswarm/draughtsman/service/notifier"
+	notifierspec "github.com/giantswarm/draughtsman/service/notifier/spec"
 	slackspec "github.com/giantswarm/draughtsman/service/slack"
 )
